+++ src/diffusers/loaders/lora.py
@@ -11,10 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import copy
-import inspect
 import os
-from pathlib import Path
 from typing import Callable, Dict, List, Optional, Union
 
 import safetensors
@@ -22,7 +19,6 @@ import torch
 from huggingface_hub import model_info
 from huggingface_hub.constants import HF_HUB_OFFLINE
 from huggingface_hub.utils import validate_hf_hub_args
-from torch import nn
 
 from ..models.modeling_utils import load_state_dict
 from ..utils import (
@@ -31,7 +27,6 @@ from ..utils import (
     convert_state_dict_to_diffusers,
     convert_state_dict_to_peft,
     convert_unet_state_dict_to_peft,
-    delete_adapter_layers,
     get_adapter_name,
     get_peft_kwargs,
     is_accelerate_available,
@@ -40,19 +35,16 @@ from ..utils import (
     logging,
     recurse_remove_peft_layers,
     scale_lora_layers,
-    set_adapter_layers,
-    set_weights_and_activate_adapters,
 )
+from .lora_base import LoraUtilsMixin
 from .lora_conversion_utils import _convert_non_diffusers_lora_to_diffusers, _maybe_map_sgm_blocks_to_diffusers
 
 
 if is_transformers_available():
-    from transformers import PreTrainedModel
-
     from ..models.lora import text_encoder_attn_modules, text_encoder_mlp_modules
 
 if is_accelerate_available():
-    from accelerate.hooks import AlignDevicesHook, CpuOffload, remove_hook_from_module
+    pass
 
 logger = logging.get_logger(__name__)
 
@@ -66,7 +58,7 @@ LORA_WEIGHT_NAME_SAFE = "pytorch_lora_weights.safetensors"
 LORA_DEPRECATION_MESSAGE = "You are using an old version of LoRA backend. This will be deprecated in the next releases in favor of PEFT make sure to install the latest PEFT and transformers packages in the future."
 
 
-class LoraLoaderMixin:
+class LoraLoaderMixin(LoraUtilsMixin):
     r"""
     Load LoRA layers into [`UNet2DConditionModel`] and
     [`CLIPTextModel`](https://huggingface.co/docs/transformers/model_doc/clip#transformers.CLIPTextModel).
@@ -74,8 +66,7 @@ class LoraLoaderMixin:
 
     text_encoder_name = TEXT_ENCODER_NAME
     unet_name = UNET_NAME
-    transformer_name = TRANSFORMER_NAME
-    num_fused_loras = 0
+    is_unet_denoiser = True
 
     def load_lora_weights(
         self, pretrained_model_name_or_path_or_dict: Union[str, Dict[str, torch.Tensor]], adapter_name=None, **kwargs
@@ -333,40 +324,74 @@ class LoraLoaderMixin:
         weight_name = targeted_files[0]
         return weight_name
 
-    @classmethod
-    def _optionally_disable_offloading(cls, _pipeline):
-        """
-        Optionally removes offloading in case the pipeline has been already sequentially offloaded to CPU.
+    def fuse_lora(
+        self,
+        fuse_unet: bool = True,
+        fuse_text_encoder: bool = True,
+        lora_scale: float = 1.0,
+        safe_fusing: bool = False,
+        adapter_names: Optional[List[str]] = None,
+    ):
+        r"""
+        Fuses the LoRA parameters into the original parameters of the corresponding blocks.
+
+        <Tip warning={true}>
+
+        This is an experimental API.
+
+        </Tip>
 
         Args:
-            _pipeline (`DiffusionPipeline`):
-                The pipeline to disable offloading for.
+            fuse_unet (`bool`, defaults to `True`): Whether to fuse the UNet LoRA parameters.
+            fuse_text_encoder (`bool`, defaults to `True`):
+                Whether to fuse the text encoder LoRA parameters. If the text encoder wasn't monkey-patched with the
+                LoRA parameters then it won't have any effect.
+            lora_scale (`float`, defaults to 1.0):
+                Controls how much to influence the outputs with the LoRA parameters.
+            safe_fusing (`bool`, defaults to `False`):
+                Whether to check fused weights for NaN values before fusing and if values are NaN not fusing them.
+            adapter_names (`List[str]`, *optional*):
+                Adapter names to be used for fusing. If nothing is passed, all active adapters will be fused.
 
-        Returns:
-            tuple:
-                A tuple indicating if `is_model_cpu_offload` or `is_sequential_cpu_offload` is True.
+        Example:
+
+        ```py
+        from diffusers import DiffusionPipeline
+        import torch
+
+        pipeline = DiffusionPipeline.from_pretrained(
+            "stabilityai/stable-diffusion-xl-base-1.0", torch_dtype=torch.float16
+        ).to("cuda")
+        pipeline.load_lora_weights("nerijs/pixel-art-xl", weight_name="pixel-art-xl.safetensors", adapter_name="pixel")
+        pipeline.fuse_lora(lora_scale=0.7)
+        ```
         """
-        is_model_cpu_offload = False
-        is_sequential_cpu_offload = False
+        super().fuse_lora(
+            fuse_denoiser=fuse_unet,
+            fuse_text_encoder=fuse_text_encoder,
+            lora_scale=lora_scale,
+            safe_fusing=safe_fusing,
+            adapter_names=adapter_names,
+        )
 
-        if _pipeline is not None and _pipeline.hf_device_map is None:
-            for _, component in _pipeline.components.items():
-                if isinstance(component, nn.Module) and hasattr(component, "_hf_hook"):
-                    if not is_model_cpu_offload:
-                        is_model_cpu_offload = isinstance(component._hf_hook, CpuOffload)
-                    if not is_sequential_cpu_offload:
-                        is_sequential_cpu_offload = (
-                            isinstance(component._hf_hook, AlignDevicesHook)
-                            or hasattr(component._hf_hook, "hooks")
-                            and isinstance(component._hf_hook.hooks[0], AlignDevicesHook)
-                        )
+    def unfuse_lora(self, unfuse_unet: bool = True, unfuse_text_encoder: bool = True):
+        r"""
+        Reverses the effect of
+        [`pipe.fuse_lora()`](https://huggingface.co/docs/diffusers/main/en/api/loaders#diffusers.loaders.LoraLoaderMixin.fuse_lora).
 
-                    logger.info(
-                        "Accelerate hooks detected. Since you have called `load_lora_weights()`, the previous hooks will be first removed. Then the LoRA parameters will be loaded and the hooks will be applied again."
-                    )
-                    remove_hook_from_module(component, recurse=is_sequential_cpu_offload)
+        <Tip warning={true}>
 
-        return (is_model_cpu_offload, is_sequential_cpu_offload)
+        This is an experimental API.
+
+        </Tip>
+
+        Args:
+            unfuse_unet (`bool`, defaults to `True`): Whether to unfuse the UNet LoRA parameters.
+            unfuse_text_encoder (`bool`, defaults to `True`):
+                Whether to unfuse the text encoder LoRA parameters. If the text encoder wasn't monkey-patched with the
+                LoRA parameters then it won't have any effect.
+        """
+        super().unfuse_lora(unfuse_denoiser=unfuse_unet, unfuse_text_encoder=unfuse_text_encoder)
 
     @classmethod
     def load_lora_into_unet(cls, state_dict, network_alphas, unet, adapter_name=None, _pipeline=None):
@@ -522,107 +547,11 @@ class LoraLoaderMixin:
                     _pipeline.enable_sequential_cpu_offload()
                 # Unsafe code />
 
-    @classmethod
-    def load_lora_into_transformer(cls, state_dict, network_alphas, transformer, adapter_name=None, _pipeline=None):
-        """
-        This will load the LoRA layers specified in `state_dict` into `transformer`.
-
-        Parameters:
-            state_dict (`dict`):
-                A standard state dict containing the lora layer parameters. The keys can either be indexed directly
-                into the unet or prefixed with an additional `unet` which can be used to distinguish between text
-                encoder lora layers.
-            network_alphas (`Dict[str, float]`):
-                See `LoRALinearLayer` for more details.
-            unet (`UNet2DConditionModel`):
-                The UNet model to load the LoRA layers into.
-            adapter_name (`str`, *optional*):
-                Adapter name to be used for referencing the loaded adapter model. If not specified, it will use
-                `default_{i}` where i is the total number of adapters being loaded.
-        """
-        from peft import LoraConfig, inject_adapter_in_model, set_peft_model_state_dict
-
-        keys = list(state_dict.keys())
-
-        transformer_keys = [k for k in keys if k.startswith(cls.transformer_name)]
-        state_dict = {
-            k.replace(f"{cls.transformer_name}.", ""): v for k, v in state_dict.items() if k in transformer_keys
-        }
-
-        if network_alphas is not None:
-            alpha_keys = [k for k in network_alphas.keys() if k.startswith(cls.transformer_name)]
-            network_alphas = {
-                k.replace(f"{cls.transformer_name}.", ""): v for k, v in network_alphas.items() if k in alpha_keys
-            }
-
-        if len(state_dict.keys()) > 0:
-            if adapter_name in getattr(transformer, "peft_config", {}):
-                raise ValueError(
-                    f"Adapter name {adapter_name} already in use in the transformer - please select a new adapter name."
-                )
-
-            rank = {}
-            for key, val in state_dict.items():
-                if "lora_B" in key:
-                    rank[key] = val.shape[1]
-
-            lora_config_kwargs = get_peft_kwargs(rank, network_alphas, state_dict)
-            if "use_dora" in lora_config_kwargs:
-                if lora_config_kwargs["use_dora"] and is_peft_version("<", "0.9.0"):
-                    raise ValueError(
-                        "You need `peft` 0.9.0 at least to use DoRA-enabled LoRAs. Please upgrade your installation of `peft`."
-                    )
-                else:
-                    lora_config_kwargs.pop("use_dora")
-            lora_config = LoraConfig(**lora_config_kwargs)
-
-            # adapter_name
-            if adapter_name is None:
-                adapter_name = get_adapter_name(transformer)
-
-            # In case the pipeline has been already offloaded to CPU - temporarily remove the hooks
-            # otherwise loading LoRA weights will lead to an error
-            is_model_cpu_offload, is_sequential_cpu_offload = cls._optionally_disable_offloading(_pipeline)
-
-            inject_adapter_in_model(lora_config, transformer, adapter_name=adapter_name)
-            incompatible_keys = set_peft_model_state_dict(transformer, state_dict, adapter_name)
-
-            if incompatible_keys is not None:
-                # check only for unexpected keys
-                unexpected_keys = getattr(incompatible_keys, "unexpected_keys", None)
-                if unexpected_keys:
-                    logger.warning(
-                        f"Loading adapter weights from state_dict led to unexpected keys not found in the model: "
-                        f" {unexpected_keys}. "
-                    )
-
-            # Offload back.
-            if is_model_cpu_offload:
-                _pipeline.enable_model_cpu_offload()
-            elif is_sequential_cpu_offload:
-                _pipeline.enable_sequential_cpu_offload()
-            # Unsafe code />
-
     @property
     def lora_scale(self) -> float:
         # property function that returns the lora scale which can be set at run time by the pipeline.
         # if _lora_scale has not been set, return 1
         return self._lora_scale if hasattr(self, "_lora_scale") else 1.0
-
-    def _remove_text_encoder_monkey_patch(self):
-        remove_method = recurse_remove_peft_layers
-        if hasattr(self, "text_encoder"):
-            remove_method(self.text_encoder)
-            # In case text encoder have no Lora attached
-            if getattr(self.text_encoder, "peft_config", None) is not None:
-                del self.text_encoder.peft_config
-                self.text_encoder._hf_peft_config_loaded = None
-
-        if hasattr(self, "text_encoder_2"):
-            remove_method(self.text_encoder_2)
-            if getattr(self.text_encoder_2, "peft_config", None) is not None:
-                del self.text_encoder_2.peft_config
-                self.text_encoder_2._hf_peft_config_loaded = None
 
     @classmethod
     def save_lora_weights(
@@ -630,7 +559,6 @@ class LoraLoaderMixin:
         save_directory: Union[str, os.PathLike],
         unet_lora_layers: Dict[str, Union[torch.nn.Module, torch.Tensor]] = None,
         text_encoder_lora_layers: Dict[str, torch.nn.Module] = None,
-        transformer_lora_layers: Dict[str, torch.nn.Module] = None,
         is_main_process: bool = True,
         weight_name: str = None,
         save_function: Callable = None,
@@ -660,24 +588,14 @@ class LoraLoaderMixin:
         """
         state_dict = {}
 
-        def pack_weights(layers, prefix):
-            layers_weights = layers.state_dict() if isinstance(layers, torch.nn.Module) else layers
-            layers_state_dict = {f"{prefix}.{module_name}": param for module_name, param in layers_weights.items()}
-            return layers_state_dict
-
-        if not (unet_lora_layers or text_encoder_lora_layers or transformer_lora_layers):
-            raise ValueError(
-                "You must pass at least one of `unet_lora_layers`, `text_encoder_lora_layers`, or `transformer_lora_layers`."
-            )
+        if not (unet_lora_layers or text_encoder_lora_layers):
+            raise ValueError("You must pass at least one of `unet_lora_layers` and `text_encoder_lora_layers`.")
 
         if unet_lora_layers:
-            state_dict.update(pack_weights(unet_lora_layers, cls.unet_name))
+            state_dict.update(cls.pack_weights(unet_lora_layers, cls.unet_name))
 
         if text_encoder_lora_layers:
-            state_dict.update(pack_weights(text_encoder_lora_layers, cls.text_encoder_name))
-
-        if transformer_lora_layers:
-            state_dict.update(pack_weights(transformer_lora_layers, "transformer"))
+            state_dict.update(cls.pack_weights(text_encoder_lora_layers, cls.text_encoder_name))
 
         # Save the model
         cls.write_lora_layers(
@@ -689,499 +607,10 @@ class LoraLoaderMixin:
             safe_serialization=safe_serialization,
         )
 
-    @staticmethod
-    def write_lora_layers(
-        state_dict: Dict[str, torch.Tensor],
-        save_directory: str,
-        is_main_process: bool,
-        weight_name: str,
-        save_function: Callable,
-        safe_serialization: bool,
-    ):
-        if os.path.isfile(save_directory):
-            logger.error(f"Provided path ({save_directory}) should be a directory, not a file")
-            return
-
-        if save_function is None:
-            if safe_serialization:
-
-                def save_function(weights, filename):
-                    return safetensors.torch.save_file(weights, filename, metadata={"format": "pt"})
-
-            else:
-                save_function = torch.save
-
-        os.makedirs(save_directory, exist_ok=True)
-
-        if weight_name is None:
-            if safe_serialization:
-                weight_name = LORA_WEIGHT_NAME_SAFE
-            else:
-                weight_name = LORA_WEIGHT_NAME
-
-        save_path = Path(save_directory, weight_name).as_posix()
-        save_function(state_dict, save_path)
-        logger.info(f"Model weights saved in {save_path}")
-
-    def unload_lora_weights(self):
-        """
-        Unloads the LoRA parameters.
-
-        Examples:
-
-        ```python
-        >>> # Assuming `pipeline` is already loaded with the LoRA parameters.
-        >>> pipeline.unload_lora_weights()
-        >>> ...
-        ```
-        """
-        if not USE_PEFT_BACKEND:
-            raise ValueError("PEFT backend is required for this method.")
-
-        unet = getattr(self, self.unet_name) if not hasattr(self, "unet") else self.unet
-        unet.unload_lora()
-
-        # Safe to call the following regardless of LoRA.
-        self._remove_text_encoder_monkey_patch()
-
-    def fuse_lora(
-        self,
-        fuse_unet: bool = True,
-        fuse_text_encoder: bool = True,
-        lora_scale: float = 1.0,
-        safe_fusing: bool = False,
-        adapter_names: Optional[List[str]] = None,
-    ):
-        r"""
-        Fuses the LoRA parameters into the original parameters of the corresponding blocks.
-
-        <Tip warning={true}>
-
-        This is an experimental API.
-
-        </Tip>
-
-        Args:
-            fuse_unet (`bool`, defaults to `True`): Whether to fuse the UNet LoRA parameters.
-            fuse_text_encoder (`bool`, defaults to `True`):
-                Whether to fuse the text encoder LoRA parameters. If the text encoder wasn't monkey-patched with the
-                LoRA parameters then it won't have any effect.
-            lora_scale (`float`, defaults to 1.0):
-                Controls how much to influence the outputs with the LoRA parameters.
-            safe_fusing (`bool`, defaults to `False`):
-                Whether to check fused weights for NaN values before fusing and if values are NaN not fusing them.
-            adapter_names (`List[str]`, *optional*):
-                Adapter names to be used for fusing. If nothing is passed, all active adapters will be fused.
-
-        Example:
-
-        ```py
-        from diffusers import DiffusionPipeline
-        import torch
-
-        pipeline = DiffusionPipeline.from_pretrained(
-            "stabilityai/stable-diffusion-xl-base-1.0", torch_dtype=torch.float16
-        ).to("cuda")
-        pipeline.load_lora_weights("nerijs/pixel-art-xl", weight_name="pixel-art-xl.safetensors", adapter_name="pixel")
-        pipeline.fuse_lora(lora_scale=0.7)
-        ```
-        """
-        from peft.tuners.tuners_utils import BaseTunerLayer
-
-        if fuse_unet or fuse_text_encoder:
-            self.num_fused_loras += 1
-            if self.num_fused_loras > 1:
-                logger.warning(
-                    "The current API is supported for operating with a single LoRA file. You are trying to load and fuse more than one LoRA which is not well-supported.",
-                )
-
-        if fuse_unet:
-            unet = getattr(self, self.unet_name) if not hasattr(self, "unet") else self.unet
-            unet.fuse_lora(lora_scale, safe_fusing=safe_fusing, adapter_names=adapter_names)
-
-        def fuse_text_encoder_lora(text_encoder, lora_scale=1.0, safe_fusing=False, adapter_names=None):
-            merge_kwargs = {"safe_merge": safe_fusing}
-
-            for module in text_encoder.modules():
-                if isinstance(module, BaseTunerLayer):
-                    if lora_scale != 1.0:
-                        module.scale_layer(lora_scale)
-
-                    # For BC with previous PEFT versions, we need to check the signature
-                    # of the `merge` method to see if it supports the `adapter_names` argument.
-                    supported_merge_kwargs = list(inspect.signature(module.merge).parameters)
-                    if "adapter_names" in supported_merge_kwargs:
-                        merge_kwargs["adapter_names"] = adapter_names
-                    elif "adapter_names" not in supported_merge_kwargs and adapter_names is not None:
-                        raise ValueError(
-                            "The `adapter_names` argument is not supported with your PEFT version. "
-                            "Please upgrade to the latest version of PEFT. `pip install -U peft`"
-                        )
-
-                    module.merge(**merge_kwargs)
-
-        if fuse_text_encoder:
-            if hasattr(self, "text_encoder"):
-                fuse_text_encoder_lora(self.text_encoder, lora_scale, safe_fusing, adapter_names=adapter_names)
-            if hasattr(self, "text_encoder_2"):
-                fuse_text_encoder_lora(self.text_encoder_2, lora_scale, safe_fusing, adapter_names=adapter_names)
-
-    def unfuse_lora(self, unfuse_unet: bool = True, unfuse_text_encoder: bool = True):
-        r"""
-        Reverses the effect of
-        [`pipe.fuse_lora()`](https://huggingface.co/docs/diffusers/main/en/api/loaders#diffusers.loaders.LoraLoaderMixin.fuse_lora).
-
-        <Tip warning={true}>
-
-        This is an experimental API.
-
-        </Tip>
-
-        Args:
-            unfuse_unet (`bool`, defaults to `True`): Whether to unfuse the UNet LoRA parameters.
-            unfuse_text_encoder (`bool`, defaults to `True`):
-                Whether to unfuse the text encoder LoRA parameters. If the text encoder wasn't monkey-patched with the
-                LoRA parameters then it won't have any effect.
-        """
-        from peft.tuners.tuners_utils import BaseTunerLayer
-
-        unet = getattr(self, self.unet_name) if not hasattr(self, "unet") else self.unet
-        if unfuse_unet:
-            for module in unet.modules():
-                if isinstance(module, BaseTunerLayer):
-                    module.unmerge()
-
-        def unfuse_text_encoder_lora(text_encoder):
-            for module in text_encoder.modules():
-                if isinstance(module, BaseTunerLayer):
-                    module.unmerge()
-
-        if unfuse_text_encoder:
-            if hasattr(self, "text_encoder"):
-                unfuse_text_encoder_lora(self.text_encoder)
-            if hasattr(self, "text_encoder_2"):
-                unfuse_text_encoder_lora(self.text_encoder_2)
-
-        self.num_fused_loras -= 1
-
-    def set_adapters_for_text_encoder(
-        self,
-        adapter_names: Union[List[str], str],
-        text_encoder: Optional["PreTrainedModel"] = None,  # noqa: F821
-        text_encoder_weights: Optional[Union[float, List[float], List[None]]] = None,
-    ):
-        """
-        Sets the adapter layers for the text encoder.
-
-        Args:
-            adapter_names (`List[str]` or `str`):
-                The names of the adapters to use.
-            text_encoder (`torch.nn.Module`, *optional*):
-                The text encoder module to set the adapter layers for. If `None`, it will try to get the `text_encoder`
-                attribute.
-            text_encoder_weights (`List[float]`, *optional*):
-                The weights to use for the text encoder. If `None`, the weights are set to `1.0` for all the adapters.
-        """
-        if not USE_PEFT_BACKEND:
-            raise ValueError("PEFT backend is required for this method.")
-
-        def process_weights(adapter_names, weights):
-            # Expand weights into a list, one entry per adapter
-            # e.g. for 2 adapters:  7 -> [7,7] ; [3, None] -> [3, None]
-            if not isinstance(weights, list):
-                weights = [weights] * len(adapter_names)
-
-            if len(adapter_names) != len(weights):
-                raise ValueError(
-                    f"Length of adapter names {len(adapter_names)} is not equal to the length of the weights {len(weights)}"
-                )
-
-            # Set None values to default of 1.0
-            # e.g. [7,7] -> [7,7] ; [3, None] -> [3,1]
-            weights = [w if w is not None else 1.0 for w in weights]
-
-            return weights
-
-        adapter_names = [adapter_names] if isinstance(adapter_names, str) else adapter_names
-        text_encoder_weights = process_weights(adapter_names, text_encoder_weights)
-        text_encoder = text_encoder or getattr(self, "text_encoder", None)
-        if text_encoder is None:
-            raise ValueError(
-                "The pipeline does not have a default `pipe.text_encoder` class. Please make sure to pass a `text_encoder` instead."
-            )
-        set_weights_and_activate_adapters(text_encoder, adapter_names, text_encoder_weights)
-
-    def disable_lora_for_text_encoder(self, text_encoder: Optional["PreTrainedModel"] = None):
-        """
-        Disables the LoRA layers for the text encoder.
-
-        Args:
-            text_encoder (`torch.nn.Module`, *optional*):
-                The text encoder module to disable the LoRA layers for. If `None`, it will try to get the
-                `text_encoder` attribute.
-        """
-        if not USE_PEFT_BACKEND:
-            raise ValueError("PEFT backend is required for this method.")
-
-        text_encoder = text_encoder or getattr(self, "text_encoder", None)
-        if text_encoder is None:
-            raise ValueError("Text Encoder not found.")
-        set_adapter_layers(text_encoder, enabled=False)
-
-    def enable_lora_for_text_encoder(self, text_encoder: Optional["PreTrainedModel"] = None):
-        """
-        Enables the LoRA layers for the text encoder.
-
-        Args:
-            text_encoder (`torch.nn.Module`, *optional*):
-                The text encoder module to enable the LoRA layers for. If `None`, it will try to get the `text_encoder`
-                attribute.
-        """
-        if not USE_PEFT_BACKEND:
-            raise ValueError("PEFT backend is required for this method.")
-        text_encoder = text_encoder or getattr(self, "text_encoder", None)
-        if text_encoder is None:
-            raise ValueError("Text Encoder not found.")
-        set_adapter_layers(self.text_encoder, enabled=True)
-
-    def set_adapters(
-        self,
-        adapter_names: Union[List[str], str],
-        adapter_weights: Optional[Union[float, Dict, List[float], List[Dict]]] = None,
-    ):
-        adapter_names = [adapter_names] if isinstance(adapter_names, str) else adapter_names
-
-        adapter_weights = copy.deepcopy(adapter_weights)
-
-        # Expand weights into a list, one entry per adapter
-        if not isinstance(adapter_weights, list):
-            adapter_weights = [adapter_weights] * len(adapter_names)
-
-        if len(adapter_names) != len(adapter_weights):
-            raise ValueError(
-                f"Length of adapter names {len(adapter_names)} is not equal to the length of the weights {len(adapter_weights)}"
-            )
-
-        # Decompose weights into weights for unet, text_encoder and text_encoder_2
-        unet_lora_weights, text_encoder_lora_weights, text_encoder_2_lora_weights = [], [], []
-
-        list_adapters = self.get_list_adapters()  # eg {"unet": ["adapter1", "adapter2"], "text_encoder": ["adapter2"]}
-        all_adapters = {
-            adapter for adapters in list_adapters.values() for adapter in adapters
-        }  # eg ["adapter1", "adapter2"]
-        invert_list_adapters = {
-            adapter: [part for part, adapters in list_adapters.items() if adapter in adapters]
-            for adapter in all_adapters
-        }  # eg {"adapter1": ["unet"], "adapter2": ["unet", "text_encoder"]}
-
-        for adapter_name, weights in zip(adapter_names, adapter_weights):
-            if isinstance(weights, dict):
-                unet_lora_weight = weights.pop("unet", None)
-                text_encoder_lora_weight = weights.pop("text_encoder", None)
-                text_encoder_2_lora_weight = weights.pop("text_encoder_2", None)
-
-                if len(weights) > 0:
-                    raise ValueError(
-                        f"Got invalid key '{weights.keys()}' in lora weight dict for adapter {adapter_name}."
-                    )
-
-                if text_encoder_2_lora_weight is not None and not hasattr(self, "text_encoder_2"):
-                    logger.warning(
-                        "Lora weight dict contains text_encoder_2 weights but will be ignored because pipeline does not have text_encoder_2."
-                    )
-
-                # warn if adapter doesn't have parts specified by adapter_weights
-                for part_weight, part_name in zip(
-                    [unet_lora_weight, text_encoder_lora_weight, text_encoder_2_lora_weight],
-                    ["unet", "text_encoder", "text_encoder_2"],
-                ):
-                    if part_weight is not None and part_name not in invert_list_adapters[adapter_name]:
-                        logger.warning(
-                            f"Lora weight dict for adapter '{adapter_name}' contains {part_name}, but this will be ignored because {adapter_name} does not contain weights for {part_name}. Valid parts for {adapter_name} are: {invert_list_adapters[adapter_name]}."
-                        )
-
-            else:
-                unet_lora_weight = weights
-                text_encoder_lora_weight = weights
-                text_encoder_2_lora_weight = weights
-
-            unet_lora_weights.append(unet_lora_weight)
-            text_encoder_lora_weights.append(text_encoder_lora_weight)
-            text_encoder_2_lora_weights.append(text_encoder_2_lora_weight)
-
-        unet = getattr(self, self.unet_name) if not hasattr(self, "unet") else self.unet
-        # Handle the UNET
-        unet.set_adapters(adapter_names, unet_lora_weights)
-
-        # Handle the Text Encoder
-        if hasattr(self, "text_encoder"):
-            self.set_adapters_for_text_encoder(adapter_names, self.text_encoder, text_encoder_lora_weights)
-        if hasattr(self, "text_encoder_2"):
-            self.set_adapters_for_text_encoder(adapter_names, self.text_encoder_2, text_encoder_2_lora_weights)
-
-    def disable_lora(self):
-        if not USE_PEFT_BACKEND:
-            raise ValueError("PEFT backend is required for this method.")
-
-        # Disable unet adapters
-        unet = getattr(self, self.unet_name) if not hasattr(self, "unet") else self.unet
-        unet.disable_lora()
-
-        # Disable text encoder adapters
-        if hasattr(self, "text_encoder"):
-            self.disable_lora_for_text_encoder(self.text_encoder)
-        if hasattr(self, "text_encoder_2"):
-            self.disable_lora_for_text_encoder(self.text_encoder_2)
-
-    def enable_lora(self):
-        if not USE_PEFT_BACKEND:
-            raise ValueError("PEFT backend is required for this method.")
-
-        # Enable unet adapters
-        unet = getattr(self, self.unet_name) if not hasattr(self, "unet") else self.unet
-        unet.enable_lora()
-
-        # Enable text encoder adapters
-        if hasattr(self, "text_encoder"):
-            self.enable_lora_for_text_encoder(self.text_encoder)
-        if hasattr(self, "text_encoder_2"):
-            self.enable_lora_for_text_encoder(self.text_encoder_2)
-
-    def delete_adapters(self, adapter_names: Union[List[str], str]):
-        """
-        Args:
-        Deletes the LoRA layers of `adapter_name` for the unet and text-encoder(s).
-            adapter_names (`Union[List[str], str]`):
-                The names of the adapter to delete. Can be a single string or a list of strings
-        """
-        if not USE_PEFT_BACKEND:
-            raise ValueError("PEFT backend is required for this method.")
-
-        if isinstance(adapter_names, str):
-            adapter_names = [adapter_names]
-
-        # Delete unet adapters
-        unet = getattr(self, self.unet_name) if not hasattr(self, "unet") else self.unet
-        unet.delete_adapters(adapter_names)
-
-        for adapter_name in adapter_names:
-            # Delete text encoder adapters
-            if hasattr(self, "text_encoder"):
-                delete_adapter_layers(self.text_encoder, adapter_name)
-            if hasattr(self, "text_encoder_2"):
-                delete_adapter_layers(self.text_encoder_2, adapter_name)
-
-    def get_active_adapters(self) -> List[str]:
-        """
-        Gets the list of the current active adapters.
-
-        Example:
-
-        ```python
-        from diffusers import DiffusionPipeline
-
-        pipeline = DiffusionPipeline.from_pretrained(
-            "stabilityai/stable-diffusion-xl-base-1.0",
-        ).to("cuda")
-        pipeline.load_lora_weights("CiroN2022/toy-face", weight_name="toy_face_sdxl.safetensors", adapter_name="toy")
-        pipeline.get_active_adapters()
-        ```
-        """
-        if not USE_PEFT_BACKEND:
-            raise ValueError(
-                "PEFT backend is required for this method. Please install the latest version of PEFT `pip install -U peft`"
-            )
-
-        from peft.tuners.tuners_utils import BaseTunerLayer
-
-        active_adapters = []
-        unet = getattr(self, self.unet_name) if not hasattr(self, "unet") else self.unet
-        for module in unet.modules():
-            if isinstance(module, BaseTunerLayer):
-                active_adapters = module.active_adapters
-                break
-
-        return active_adapters
-
-    def get_list_adapters(self) -> Dict[str, List[str]]:
-        """
-        Gets the current list of all available adapters in the pipeline.
-        """
-        if not USE_PEFT_BACKEND:
-            raise ValueError(
-                "PEFT backend is required for this method. Please install the latest version of PEFT `pip install -U peft`"
-            )
-
-        set_adapters = {}
-
-        if hasattr(self, "text_encoder") and hasattr(self.text_encoder, "peft_config"):
-            set_adapters["text_encoder"] = list(self.text_encoder.peft_config.keys())
-
-        if hasattr(self, "text_encoder_2") and hasattr(self.text_encoder_2, "peft_config"):
-            set_adapters["text_encoder_2"] = list(self.text_encoder_2.peft_config.keys())
-
-        unet = getattr(self, self.unet_name) if not hasattr(self, "unet") else self.unet
-        if hasattr(self, self.unet_name) and hasattr(unet, "peft_config"):
-            set_adapters[self.unet_name] = list(self.unet.peft_config.keys())
-
-        return set_adapters
-
-    def set_lora_device(self, adapter_names: List[str], device: Union[torch.device, str, int]) -> None:
-        """
-        Moves the LoRAs listed in `adapter_names` to a target device. Useful for offloading the LoRA to the CPU in case
-        you want to load multiple adapters and free some GPU memory.
-
-        Args:
-            adapter_names (`List[str]`):
-                List of adapters to send device to.
-            device (`Union[torch.device, str, int]`):
-                Device to send the adapters to. Can be either a torch device, a str or an integer.
-        """
-        if not USE_PEFT_BACKEND:
-            raise ValueError("PEFT backend is required for this method.")
-
-        from peft.tuners.tuners_utils import BaseTunerLayer
-
-        # Handle the UNET
-        unet = getattr(self, self.unet_name) if not hasattr(self, "unet") else self.unet
-        for unet_module in unet.modules():
-            if isinstance(unet_module, BaseTunerLayer):
-                for adapter_name in adapter_names:
-                    unet_module.lora_A[adapter_name].to(device)
-                    unet_module.lora_B[adapter_name].to(device)
-                    # this is a param, not a module, so device placement is not in-place -> re-assign
-                    if hasattr(unet_module, "lora_magnitude_vector") and unet_module.lora_magnitude_vector is not None:
-                        unet_module.lora_magnitude_vector[adapter_name] = unet_module.lora_magnitude_vector[
-                            adapter_name
-                        ].to(device)
-
-        # Handle the text encoder
-        modules_to_process = []
-        if hasattr(self, "text_encoder"):
-            modules_to_process.append(self.text_encoder)
-
-        if hasattr(self, "text_encoder_2"):
-            modules_to_process.append(self.text_encoder_2)
-
-        for text_encoder in modules_to_process:
-            # loop over submodules
-            for text_encoder_module in text_encoder.modules():
-                if isinstance(text_encoder_module, BaseTunerLayer):
-                    for adapter_name in adapter_names:
-                        text_encoder_module.lora_A[adapter_name].to(device)
-                        text_encoder_module.lora_B[adapter_name].to(device)
-                        # this is a param, not a module, so device placement is not in-place -> re-assign
-                        if (
-                            hasattr(text_encoder_module, "lora_magnitude_vector")
-                            and text_encoder_module.lora_magnitude_vector is not None
-                        ):
-                            text_encoder_module.lora_magnitude_vector[
-                                adapter_name
-                            ] = text_encoder_module.lora_magnitude_vector[adapter_name].to(device)
-
 
 class StableDiffusionXLLoraLoaderMixin(LoraLoaderMixin):
-    """This class overrides `LoraLoaderMixin` with LoRA loading/saving code that's specific to SDXL"""
+    """This class overrides `LoraLoaderMixin` with LoRA loading/saving code that's specific
+    to SDXL ([`StableDiffusionXLPipeline`].)"""
 
     # Override to properly handle the loading and unloading of the additional text encoder.
     def load_lora_weights(
@@ -1300,24 +729,19 @@ class StableDiffusionXLLoraLoaderMixin(LoraLoaderMixin):
         """
         state_dict = {}
 
-        def pack_weights(layers, prefix):
-            layers_weights = layers.state_dict() if isinstance(layers, torch.nn.Module) else layers
-            layers_state_dict = {f"{prefix}.{module_name}": param for module_name, param in layers_weights.items()}
-            return layers_state_dict
-
         if not (unet_lora_layers or text_encoder_lora_layers or text_encoder_2_lora_layers):
             raise ValueError(
                 "You must pass at least one of `unet_lora_layers`, `text_encoder_lora_layers` or `text_encoder_2_lora_layers`."
             )
 
         if unet_lora_layers:
-            state_dict.update(pack_weights(unet_lora_layers, "unet"))
+            state_dict.update(cls.pack_weights(unet_lora_layers, "unet"))
 
         if text_encoder_lora_layers:
-            state_dict.update(pack_weights(text_encoder_lora_layers, "text_encoder"))
+            state_dict.update(cls.pack_weights(text_encoder_lora_layers, "text_encoder"))
 
         if text_encoder_2_lora_layers:
-            state_dict.update(pack_weights(text_encoder_2_lora_layers, "text_encoder_2"))
+            state_dict.update(cls.pack_weights(text_encoder_2_lora_layers, "text_encoder_2"))
 
         cls.write_lora_layers(
             state_dict=state_dict,
@@ -1341,13 +765,82 @@ class StableDiffusionXLLoraLoaderMixin(LoraLoaderMixin):
             self.text_encoder_2._hf_peft_config_loaded = None
 
 
-class SD3LoraLoaderMixin:
+class SD3LoraLoaderMixin(LoraUtilsMixin):
     r"""
-    Load LoRA layers into [`SD3Transformer2DModel`].
+    Load LoRA layers into [`SD3Transformer2DModel`]. Specific to [`StableDiffusion3Pipeline`].
     """
 
     transformer_name = TRANSFORMER_NAME
-    num_fused_loras = 0
+    is_transformer_denoiser = True
+
+    def fuse_lora(
+        self,
+        fuse_transformer: bool = True,
+        lora_scale: float = 1.0,
+        safe_fusing: bool = False,
+        adapter_names: Optional[List[str]] = None,
+    ):
+        r"""
+        Fuses the LoRA parameters into the original parameters of the corresponding blocks.
+
+        <Tip warning={true}>
+
+        This is an experimental API.
+
+        </Tip>
+
+        Args:
+            fuse_transformer (`bool`, defaults to `True`): Whether to fuse the transformer LoRA parameters.
+            lora_scale (`float`, defaults to 1.0):
+                Controls how much to influence the outputs with the LoRA parameters.
+            safe_fusing (`bool`, defaults to `False`):
+                Whether to check fused weights for NaN values before fusing and if values are NaN not fusing them.
+            adapter_names (`List[str]`, *optional*):
+                Adapter names to be used for fusing. If nothing is passed, all active adapters will be fused.
+
+        Example:
+
+        ```py
+        from diffusers import DiffusionPipeline
+        import torch
+
+        pipeline = DiffusionPipeline.from_pretrained(
+            "stabilityai/stable-diffusion-3-medium-diffusers", torch_dtype=torch.float16
+        ).to("cuda")
+        pipeline.load_lora_weights(
+            "nerijs/pixel-art-medium-128-v0.1",
+            weight_name="pixel-art-medium-128-v0.1.safetensors",
+            adapter_name="pixel",
+        )
+        pipeline.fuse_lora(lora_scale=0.7)
+        ```
+        """
+        # Hard-code the `fuse_text_encoder` to False for now. We can change it later once text encoder LoRA support
+        # has been realized (TODO).
+        super().fuse_lora(
+            fuse_denoiser=fuse_transformer,
+            fuse_text_encoder=False,
+            lora_scale=lora_scale,
+            safe_fusing=safe_fusing,
+            adapter_names=adapter_names,
+        )
+
+    def unfuse_lora(self, unfuse_transformer: bool = True):
+        r"""
+        Reverses the effect of
+        [`pipe.fuse_lora()`](https://huggingface.co/docs/diffusers/main/en/api/loaders#diffusers.loaders.LoraLoaderMixin.fuse_lora).
+
+        <Tip warning={true}>
+
+        This is an experimental API.
+
+        </Tip>
+
+        Args:
+            unfuse_transformer (`bool`, defaults to `True`): Whether to unfuse the transformer LoRA parameters.
+        """
+        # Hard-code `unfuse_text_encoder` to False for now (TODO).
+        super().unfuse_lora(unfuse_denoiser=unfuse_transformer, unfuse_text_encoder=False)
 
     def load_lora_weights(
         self, pretrained_model_name_or_path_or_dict: Union[str, Dict[str, torch.Tensor]], adapter_name=None, **kwargs
@@ -1627,16 +1120,11 @@ class SD3LoraLoaderMixin:
         """
         state_dict = {}
 
-        def pack_weights(layers, prefix):
-            layers_weights = layers.state_dict() if isinstance(layers, torch.nn.Module) else layers
-            layers_state_dict = {f"{prefix}.{module_name}": param for module_name, param in layers_weights.items()}
-            return layers_state_dict
-
         if not transformer_lora_layers:
             raise ValueError("You must pass `transformer_lora_layers`.")
 
         if transformer_lora_layers:
-            state_dict.update(pack_weights(transformer_lora_layers, cls.transformer_name))
+            state_dict.update(cls.pack_weights(transformer_lora_layers, cls.transformer_name))
 
         # Save the model
         cls.write_lora_layers(
@@ -1648,164 +1136,150 @@ class SD3LoraLoaderMixin:
             safe_serialization=safe_serialization,
         )
 
-    @staticmethod
-    def write_lora_layers(
-        state_dict: Dict[str, torch.Tensor],
-        save_directory: str,
-        is_main_process: bool,
-        weight_name: str,
-        save_function: Callable,
-        safe_serialization: bool,
-    ):
-        if os.path.isfile(save_directory):
-            logger.error(f"Provided path ({save_directory}) should be a directory, not a file")
-            return
 
-        if save_function is None:
-            if safe_serialization:
-
-                def save_function(weights, filename):
-                    return safetensors.torch.save_file(weights, filename, metadata={"format": "pt"})
-
-            else:
-                save_function = torch.save
-
-        os.makedirs(save_directory, exist_ok=True)
-
-        if weight_name is None:
-            if safe_serialization:
-                weight_name = LORA_WEIGHT_NAME_SAFE
-            else:
-                weight_name = LORA_WEIGHT_NAME
-
-        save_path = Path(save_directory, weight_name).as_posix()
-        save_function(state_dict, save_path)
-        logger.info(f"Model weights saved in {save_path}")
-
-    def unload_lora_weights(self):
-        """
-        Unloads the LoRA parameters.
-
-        Examples:
-
-        ```python
-        >>> # Assuming `pipeline` is already loaded with the LoRA parameters.
-        >>> pipeline.unload_lora_weights()
-        >>> ...
-        ```
-        """
-        transformer = getattr(self, self.transformer_name) if not hasattr(self, "transformer") else self.transformer
-        recurse_remove_peft_layers(transformer)
-        if hasattr(transformer, "peft_config"):
-            del transformer.peft_config
+class AmusedLoraLoaderMixin(LoraLoaderMixin):
+    is_unet_denoiser = False
+    is_transformer_denoiser = True
+    transformer_name = TRANSFORMER_NAME
 
     @classmethod
-    # Copied from diffusers.loaders.lora.LoraLoaderMixin._optionally_disable_offloading
-    def _optionally_disable_offloading(cls, _pipeline):
+    def load_lora_into_unet(cls, **kwargs):
+        raise NotImplementedError("`load_lora_into_unet()` not implemented for `AmusedLoraLoaderMixin`.")
+
+    @classmethod
+    def load_lora_into_transformer(cls, state_dict, network_alphas, transformer, adapter_name=None, _pipeline=None):
         """
-        Optionally removes offloading in case the pipeline has been already sequentially offloaded to CPU.
+        This will load the LoRA layers specified in `state_dict` into `transformer`.
 
-        Args:
-            _pipeline (`DiffusionPipeline`):
-                The pipeline to disable offloading for.
-
-        Returns:
-            tuple:
-                A tuple indicating if `is_model_cpu_offload` or `is_sequential_cpu_offload` is True.
+        Parameters:
+            state_dict (`dict`):
+                A standard state dict containing the lora layer parameters. The keys can either be indexed directly
+                into the unet or prefixed with an additional `unet` which can be used to distinguish between text
+                encoder lora layers.
+            network_alphas (`Dict[str, float]`):
+                See `LoRALinearLayer` for more details.
+            unet (`UNet2DConditionModel`):
+                The UNet model to load the LoRA layers into.
+            adapter_name (`str`, *optional*):
+                Adapter name to be used for referencing the loaded adapter model. If not specified, it will use
+                `default_{i}` where i is the total number of adapters being loaded.
         """
-        is_model_cpu_offload = False
-        is_sequential_cpu_offload = False
+        if not USE_PEFT_BACKEND:
+            raise ValueError("PEFT backend is required for this method.")
 
-        if _pipeline is not None and _pipeline.hf_device_map is None:
-            for _, component in _pipeline.components.items():
-                if isinstance(component, nn.Module) and hasattr(component, "_hf_hook"):
-                    if not is_model_cpu_offload:
-                        is_model_cpu_offload = isinstance(component._hf_hook, CpuOffload)
-                    if not is_sequential_cpu_offload:
-                        is_sequential_cpu_offload = (
-                            isinstance(component._hf_hook, AlignDevicesHook)
-                            or hasattr(component._hf_hook, "hooks")
-                            and isinstance(component._hf_hook.hooks[0], AlignDevicesHook)
-                        )
+        from peft import LoraConfig, inject_adapter_in_model, set_peft_model_state_dict
 
-                    logger.info(
-                        "Accelerate hooks detected. Since you have called `load_lora_weights()`, the previous hooks will be first removed. Then the LoRA parameters will be loaded and the hooks will be applied again."
+        keys = list(state_dict.keys())
+
+        transformer_keys = [k for k in keys if k.startswith(cls.transformer_name)]
+        state_dict = {
+            k.replace(f"{cls.transformer_name}.", ""): v for k, v in state_dict.items() if k in transformer_keys
+        }
+
+        if network_alphas is not None:
+            alpha_keys = [k for k in network_alphas.keys() if k.startswith(cls.transformer_name)]
+            network_alphas = {
+                k.replace(f"{cls.transformer_name}.", ""): v for k, v in network_alphas.items() if k in alpha_keys
+            }
+
+        if len(state_dict.keys()) > 0:
+            if adapter_name in getattr(transformer, "peft_config", {}):
+                raise ValueError(
+                    f"Adapter name {adapter_name} already in use in the transformer - please select a new adapter name."
+                )
+
+            rank = {}
+            for key, val in state_dict.items():
+                if "lora_B" in key:
+                    rank[key] = val.shape[1]
+
+            lora_config_kwargs = get_peft_kwargs(rank, network_alphas, state_dict)
+            if "use_dora" in lora_config_kwargs:
+                if lora_config_kwargs["use_dora"] and is_peft_version("<", "0.9.0"):
+                    raise ValueError(
+                        "You need `peft` 0.9.0 at least to use DoRA-enabled LoRAs. Please upgrade your installation of `peft`."
                     )
-                    remove_hook_from_module(component, recurse=is_sequential_cpu_offload)
+                else:
+                    lora_config_kwargs.pop("use_dora")
+            lora_config = LoraConfig(**lora_config_kwargs)
 
-        return (is_model_cpu_offload, is_sequential_cpu_offload)
+            # adapter_name
+            if adapter_name is None:
+                adapter_name = get_adapter_name(transformer)
 
-    def fuse_lora(
-        self,
-        fuse_transformer: bool = True,
-        lora_scale: float = 1.0,
-        safe_fusing: bool = False,
-        adapter_names: Optional[List[str]] = None,
+            # In case the pipeline has been already offloaded to CPU - temporarily remove the hooks
+            # otherwise loading LoRA weights will lead to an error
+            is_model_cpu_offload, is_sequential_cpu_offload = cls._optionally_disable_offloading(_pipeline)
+
+            inject_adapter_in_model(lora_config, transformer, adapter_name=adapter_name)
+            incompatible_keys = set_peft_model_state_dict(transformer, state_dict, adapter_name)
+
+            if incompatible_keys is not None:
+                # check only for unexpected keys
+                unexpected_keys = getattr(incompatible_keys, "unexpected_keys", None)
+                if unexpected_keys:
+                    logger.warning(
+                        f"Loading adapter weights from state_dict led to unexpected keys not found in the model: "
+                        f" {unexpected_keys}. "
+                    )
+
+            # Offload back.
+            if is_model_cpu_offload:
+                _pipeline.enable_model_cpu_offload()
+            elif is_sequential_cpu_offload:
+                _pipeline.enable_sequential_cpu_offload()
+            # Unsafe code />
+
+    @classmethod
+    def save_lora_weights(
+        cls,
+        save_directory: Union[str, os.PathLike],
+        text_encoder_lora_layers: Dict[str, torch.nn.Module] = None,
+        transformer_lora_layers: Dict[str, torch.nn.Module] = None,
+        is_main_process: bool = True,
+        weight_name: str = None,
+        save_function: Callable = None,
+        safe_serialization: bool = True,
     ):
         r"""
-        Fuses the LoRA parameters into the original parameters of the corresponding blocks.
+        Save the LoRA parameters corresponding to the UNet and text encoder.
 
-        <Tip warning={true}>
+        Arguments:
+            save_directory (`str` or `os.PathLike`):
+                Directory to save LoRA parameters to. Will be created if it doesn't exist.
+            unet_lora_layers (`Dict[str, torch.nn.Module]` or `Dict[str, torch.Tensor]`):
+                State dict of the LoRA layers corresponding to the `unet`.
+            text_encoder_lora_layers (`Dict[str, torch.nn.Module]` or `Dict[str, torch.Tensor]`):
+                State dict of the LoRA layers corresponding to the `text_encoder`. Must explicitly pass the text
+                encoder LoRA state dict because it comes from 🤗 Transformers.
+            is_main_process (`bool`, *optional*, defaults to `True`):
+                Whether the process calling this is the main process or not. Useful during distributed training and you
+                need to call this function on all processes. In this case, set `is_main_process=True` only on the main
+                process to avoid race conditions.
+            save_function (`Callable`):
+                The function to use to save the state dictionary. Useful during distributed training when you need to
+                replace `torch.save` with another method. Can be configured with the environment variable
+                `DIFFUSERS_SAVE_MODE`.
+            safe_serialization (`bool`, *optional*, defaults to `True`):
+                Whether to save the model using `safetensors` or the traditional PyTorch way with `pickle`.
+        """
+        state_dict = {}
 
-        This is an experimental API.
+        if not (transformer_lora_layers or text_encoder_lora_layers):
+            raise ValueError("You must pass at least one of `transformer_lora_layers` or `text_encoder_lora_layers`.")
 
-        </Tip>
+        if transformer_lora_layers:
+            state_dict.update(cls.pack_weights(transformer_lora_layers, cls.transformer_name))
 
-        Args:
-            fuse_transformer (`bool`, defaults to `True`): Whether to fuse the transformer LoRA parameters.
-            lora_scale (`float`, defaults to 1.0):
-                Controls how much to influence the outputs with the LoRA parameters.
-            safe_fusing (`bool`, defaults to `False`):
-                Whether to check fused weights for NaN values before fusing and if values are NaN not fusing them.
-            adapter_names (`List[str]`, *optional*):
-                Adapter names to be used for fusing. If nothing is passed, all active adapters will be fused.
+        if text_encoder_lora_layers:
+            state_dict.update(cls.pack_weights(text_encoder_lora_layers, cls.text_encoder_name))
 
-        Example:
-
-        ```py
-        from diffusers import DiffusionPipeline
-        import torch
-
-        pipeline = DiffusionPipeline.from_pretrained(
-            "stabilityai/stable-diffusion-3-medium-diffusers", torch_dtype=torch.float16
-        ).to("cuda")
-        pipeline.load_lora_weights(
-            "nerijs/pixel-art-medium-128-v0.1",
-            weight_name="pixel-art-medium-128-v0.1.safetensors",
-            adapter_name="pixel",
+        # Save the model
+        cls.write_lora_layers(
+            state_dict=state_dict,
+            save_directory=save_directory,
+            is_main_process=is_main_process,
+            weight_name=weight_name,
+            save_function=save_function,
+            safe_serialization=safe_serialization,
         )
-        pipeline.fuse_lora(lora_scale=0.7)
-        ```
-        """
-        if fuse_transformer:
-            self.num_fused_loras += 1
-
-        if fuse_transformer:
-            transformer = (
-                getattr(self, self.transformer_name) if not hasattr(self, "transformer") else self.transformer
-            )
-            transformer.fuse_lora(lora_scale, safe_fusing=safe_fusing, adapter_names=adapter_names)
-
-    def unfuse_lora(self, unfuse_transformer: bool = True):
-        r"""
-        Reverses the effect of
-        [`pipe.fuse_lora()`](https://huggingface.co/docs/diffusers/main/en/api/loaders#diffusers.loaders.LoraLoaderMixin.fuse_lora).
-
-        <Tip warning={true}>
-
-        This is an experimental API.
-
-        </Tip>
-
-        Args:
-            unfuse_transformer (`bool`, defaults to `True`): Whether to unfuse the transformer LoRA parameters.
-        """
-        from peft.tuners.tuners_utils import BaseTunerLayer
-
-        transformer = getattr(self, self.transformer_name) if not hasattr(self, "transformer") else self.transformer
-        if unfuse_transformer:
-            for module in transformer.modules():
-                if isinstance(module, BaseTunerLayer):
-                    module.unmerge()
-
-        self.num_fused_loras -= 1
